+++ Menu/Menu.hpp
@@ -14,9 +14,9 @@ class Menu {
 
 public:
     explicit Menu(sf::RenderWindow * win);
-    int getState(); // returns number of level, if user intends to run some,
-                    // or 0 if user intends to run "standard mode"
-                    // or -1 if we`r still showing menu
+    size_t getState(); // returns number of level, if user intends to run some,
+                       // or 0 if user intends to run "standard mode"
+                       // or -1 if we`r still showing menu
     void open() { state = -1; };
     void update();
